chore: move controller into `Reminders` subfolder
