+++ Backend/src/Sdk/Core.Api.Sdk/CoreApiRemindersClient.cs
@@ -1,8 +1,8 @@
 using System.Net.Http.Json;
 using Mapster;
-using ApiCreateReminderRequest = Reminders.Api.Controllers.CreateReminderRequest;
-using ApiCreateReminderResponse = Reminders.Api.Controllers.CreateReminderResponse;
-using ApiGetReminderByIdResponse = Reminders.Api.Controllers.GetReminderByIdResponse;
+using ApiCreateReminderRequest = Reminders.Api.Controllers.Reminders.CreateReminderRequest;
+using ApiCreateReminderResponse = Reminders.Api.Controllers.Reminders.CreateReminderResponse;
+using ApiGetReminderByIdResponse = Reminders.Api.Controllers.Reminders.GetReminderByIdResponse;
 using SdkReminderId = Core.Api.Sdk.Contracts.Reminders.ValueObjects.ReminderId;
 using SdkCreateReminderRequest = Core.Api.Sdk.Contracts.Reminders.Commands.CreateReminderRequest;
 using SdkCreateReminderResponse = Core.Api.Sdk.Contracts.Reminders.Commands.CreateReminderResponse;
